Hunterize J.D. Koftinoff Software, Ltd's experimental C++ MIDI Library
- Overhaul CMake files to allow for easier importing of library into other CMake based projects.
- Add version information and pre-processed version header file to expose the library version in code.
- Change building of utility and test applications to be optionally disabled by default.
- Add custom CMake source file source groups.
- Remove unnecessary pre-generated IDE project files.
- Add gitignore file to prevent undesireable files from being committed.
- Fix various build errors and precision loss compile warnings.

+++ src/jdksmidi_advancedsequencer.cpp
@@ -323,7 +323,7 @@ void AdvancedSequencer::GoToMeasure( int measure, int beat )
     // figure out which warp item we use
     // try warp to the last warp point BEFORE the
     // requested measure
-    unsigned int warp_to_item = (unsigned int)measure / MEASURES_PER_WARP;
+    size_t warp_to_item = (unsigned int)measure / MEASURES_PER_WARP;
 
     if ( warp_to_item >= warp_positions.size() )
         warp_to_item = warp_positions.size() - 1;
@@ -567,7 +567,7 @@ unsigned long AdvancedSequencer::GetCurrentTimeInMs() const
     }
     else
     {
-        return seq->GetCurrentTimeInMs();
+        return static_cast<unsigned long>(seq->GetCurrentTimeInMs());
     }
 }
